[feat] entity와 repository 추가

+++ src/main/java/com/fab/banggabgo/repository/ArticleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<Article,Long>, ArticleRepositoryCustom {
+public interface ArticleRepository extends JpaRepository<Article,Integer>, ArticleRepositoryCustom {
 
   int countByUserAndIsDeletedFalseAndIsRecruitingTrue(User user);
 }
